pyproject.toml
[project]
name = "helix"
version = "0.1.0"
description = "Autonomous AI research framework"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tomli>=2.0; python_version < '3.11'",
    "tomli-w>=1.0",
    "typer>=0.12",
    "rich>=13.0",
    "httpx>=0.27",
]

[project.scripts]
helix = "helix.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=8.0"]
